Moved record update related function into modules/records/utils. Concatenated get_all_updated_records_since with update_record_info as a cli comand update_record_info_since. update_record_info is now a normal python function, and cli_update_record_info is the comand line callable version.
